Room System: Added basic room system
The room system currently consists of a single type of rooms, each has four doors, one for each direction.
Once the player is mid-way to another room the camera switches to the other room, the room is disabled (with a delay) and all the previous neighbors are entered in to a pool of unused rooms and then taken by the new room to be it's rooms (if there will be more then one type of room, these pools will become more relevant as they will not simply empty all the time).
Each room has a permanent RoomNode (once it or it's neighbor is first visited) to save the properties of this specific room.