Added Enemy placeholder
Enemies in every room are chosen randomly such that the sum of their ranks would be equal to the rank of their ranks.
Enemies in every room will restart (but stay the same enemies) when you return to a room you have already entered.
"Enemies" are disabled on trigger enter.